Improve margin calculations
Setting the \hoffset and \voffset to -1in is a problem with packages
such as pdfpages, because the PDFs get embedded at an offset that is
very difficult to correct. By instead allowing for the 1in offset in
the calculations, people can use pdfpages more easily (e.g., including
publications in appendices).